requirements.txt
opencv-python
ultralytics
numpy
scikit-learn
Pillow
torch
torchvision
# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0